requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai
chromadb
sentence-transformers
docker
requests
pytest
